+++ src/plugins/auth/methods/policy/index.ts
@@ -4,27 +4,33 @@ import { policy as policyPolicies } from '@plugins/templates/policies';
 
 import { NotAllowed } from '@plugins/errors';
 
-import type { IPolicyDoc, IPolicyModel, IUserDoc } from '@models/types';
+import type {
+  IPolicyDoc,
+  IPolicyLeanDoc,
+  IPolicyModel,
+  IUserDoc,
+} from '@models/types';
+import type { IEditDatabaseResult } from '@plugins/auth/helpers/types';
 
 /**
  * Edit Policies in the Database
  *
  * @param {IUserDoc} admin - Admin User to Perform the Action
- * @param {IPolicyDoc} data - Data to be Modified
+ * @param {IPolicyLeanDoc} data - Data to be Modified
  * @param {Partial<IPolicyDoc>} modifiedData - Modified Object
- * @returns {Promise<boolean>} - true/false
+ * @returns {Promise<IEditDatabaseResult>} - IEditDatabaseResult
  */
 function edit(
   admin: IUserDoc,
-  data: IPolicyDoc,
+  data: IPolicyDoc | IPolicyLeanDoc,
   modifiedData: Partial<IPolicyDoc>,
-): Promise<boolean> {
+): Promise<IEditDatabaseResult> {
   const policies = [policyPolicies.edit];
   const { code, ...otherData } = modifiedData;
   if (data.code === code || code === undefined) {
     return editDatainDatabase<IPolicyDoc, IPolicyModel>(
       Policies,
-      data,
+      data._id,
       { $set: otherData },
       admin,
       policies,