+++ src/plugins/server/generators/index.ts
@@ -1,2 +1,4 @@
-export { EndpointGenerator } from './endpoint-server';
-export { SetupGenerator } from './setup';
+export * from './health-checker';
+export * from './endpoints-generator';
+export * from './setup-route-generator';
+export * from './database-handler';