+++ src/plugins/server/responses/error.ts
@@ -10,12 +10,14 @@ import type { IErrorResponse } from '@plugins/server/types';
  *
  * @param {Response} res - Express Response Object
  * @param {BaseError} error - Error Class which Extends BaseError from Plugins
+ * @param {any} data - Data if any to be passed with the error response
  */
-function errorResponse(res: Response, error: BaseError): void {
+function errorResponse(res: Response, error: BaseError, data?: unknown): void {
   const errorResponse: IErrorResponse = {
     status: error.status,
     errorname: error.errorname,
     message: error.message,
+    data,
   };
   sendResponse<IErrorResponse>(res, error.status, errorResponse);
 }
@@ -25,17 +27,22 @@ function errorResponse(res: Response, error: BaseError): void {
  *
  * @param {Response} res - Express Response Object
  * @param {any} error - error object from catch block
+ * @param {any} data - Data if any to be passed with the error response
  */
-export function errorResponseHandler(res: Response, error: unknown): void {
+export function errorResponseHandler(
+  res: Response,
+  error: unknown,
+  data?: unknown,
+): void {
   if (error instanceof BaseError) {
-    errorResponse(res, error);
+    errorResponse(res, error, data);
   } else if (error instanceof Error) {
     const serverError = new InternalServerError(error.message, error.name);
-    errorResponse(res, serverError);
+    errorResponse(res, serverError, data);
   } else {
     const serverError = new InternalServerError(
       'Unknown Error in the Server, Try Again later',
     );
-    errorResponse(res, serverError);
+    errorResponse(res, serverError, data);
   }
 }