+++ src/plugins/server/types.d.ts
@@ -1,3 +1,5 @@
+import { IEditDatabaseResult } from '@plugins/auth/helpers/types';
+
 export interface Routes {
   get: string[];
   post: string[];
@@ -15,4 +17,12 @@ export interface IErrorResponse {
   status: number;
   errorname: string;
   message: string;
+  data?: unknown;
+}
+
+export interface IEditDatabaseResponse {
+  recordsUpdated: IEditDatabaseResult[];
+  totalRecordsUpdated: number;
+  recordsNotUpdated: IEditDatabaseResult[];
+  totalRecordsNotUpdated: number;
 }